TW-1316: Persona external ads. Refactor. + src/content-scripts/replace-ads
