+++ src/content-scripts/replace-ads/ads-actions/helpers.ts
