+++ src/content-scripts/replace-ads/ads-actions/types.ts
