+++ src/content-scripts/replace-ads/ads-viewes/make-hypelab-ad.ts
