+++ src/content-scripts/replace-ads/ads-viewes/types.ts
