+++ src/content-scripts/replace-ads/utils.ts
@@ -1,0 +1,5 @@
+export const overrideElementStyles = (element: HTMLElement, overrides: StringRecord) => {
+  for (const stylePropName in overrides) {
+    element.style.setProperty(stylePropName, overrides[stylePropName]);
+  }
+};
